requirements.txt
DjangoRangeMiddleware